feat(l1-migration): rename replay to validate
